cicd/GDriveOperations/requirements.txt
google-api-python-client == 2.77.0
google-auth-httplib2 == 0.1.0
google-auth-oauthlib == 1.0.0
